+++ packages/core/mashroom/src/config/mashroom_default_config.ts
@@ -6,6 +6,9 @@ import type {MashroomServerConfig} from '../../type-definitions';
 const config: MashroomServerConfig = {
     name: 'Mashroom Server',
     port: 5050,
+    httpsPort: null,
+    tlsOptions: null,
+    enableHttp2: false,
     xPowerByHeader: 'Mashroom Server',
     serverRootFolder: '.',
     tmpFolder: os.tmpdir(),