+++ packages/plugin-packages/mashroom-security-provider-ldap/type-definitions/index.ts
@@ -23,22 +23,3 @@ export type GroupToRoleMapping = {
 export type UserToRoleMapping = {
     [userName: string]: Array<string>;
 }
-
-export type TlsOptions = {
-    ca?: string | Array<string>;
-    cert: string | Array<string>;
-    ciphers?: string;
-    clientCertEngine?: string;
-    crl?: string | Array<string>;
-    dhparam?: string;
-    ecdhCurve?: string;
-    honorCipherOrder?: boolean;
-    key?: string | Array<string>;
-    maxVersion?: string;
-    minVersion?: string;
-    passphrase?: string;
-    pfx?: string | Array<string>;
-    secureOptions?: number;
-    secureProtocol?: string;
-    rejectUnauthorized?: boolean;
-}